refactor Filter.Registry
